README updates: versa_ecp5, initrd dev file order
With recent yosys (#9cdc6b5), trellis (#210a0a7), and nextpnr
(#f0e30ab), the design reliably comes in at or below 99%
utilization, and with recent litex-boards (#8fb8005) the Ethernet
interface works for both tftp-boot under the bios, and under Linux.

Also contains a minor (cosmetic) reordering of the creation of dev
files for the initrd (software) portion of the instructions.